metrics are outside handler
