+++ internal/clientconn/conn.go
@@ -24,10 +24,13 @@ import (
 	"time"
 
 	"github.com/pmezard/go-difflib/difflib"
+	"github.com/prometheus/client_golang/prometheus"
 	"go.uber.org/zap"
 
 	"github.com/FerretDB/FerretDB/internal/handlers/common"
 	"github.com/FerretDB/FerretDB/internal/handlers/proxy"
+	"github.com/FerretDB/FerretDB/internal/types"
+	"github.com/FerretDB/FerretDB/internal/util/lazyerrors"
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
 
@@ -56,6 +59,7 @@ type conn struct {
 	mode    Mode
 	l       *zap.SugaredLogger
 	h       common.Handler
+	m       *ListenerMetrics
 	proxy   *proxy.Handler
 }
 
@@ -65,6 +69,7 @@ type newConnOpts struct {
 	mode      Mode
 	l         *zap.Logger
 	handler   common.Handler
+	m         *ListenerMetrics
 	proxyAddr string
 	startTime time.Time
 }
@@ -73,6 +78,10 @@ type newConnOpts struct {
 func newConn(opts *newConnOpts) (*conn, error) {
 	if opts.handler == nil {
 		panic("handler required")
+	}
+
+	if opts.m == nil {
+		panic("metrics required")
 	}
 
 	prefix := fmt.Sprintf("// %s -> %s ", opts.netConn.RemoteAddr(), opts.netConn.LocalAddr())
@@ -91,6 +100,7 @@ func newConn(opts *newConnOpts) (*conn, error) {
 		mode:    opts.mode,
 		l:       l.Sugar(),
 		h:       opts.handler,
+		m:       opts.m,
 		proxy:   p,
 	}, nil
 }
@@ -163,13 +173,7 @@ func (c *conn) run(ctx context.Context) (err error) {
 		var resBody wire.MsgBody
 		var resCloseConn bool
 		if c.mode != ProxyMode {
-			resHeader, resBody, resCloseConn = c.h.Handle(ctx, reqHeader, reqBody)
-
-			// do not spend time dumping if we are not going to log it
-			if c.l.Desugar().Core().Enabled(zap.DebugLevel) {
-				c.l.Debugf("Response header: %s", resHeader)
-				c.l.Debugf("Response message:\n%s\n\n\n", resBody)
-			}
+			resHeader, resBody, resCloseConn = c.route(c.h, ctx, reqHeader, reqBody)
 		}
 
 		// send request to proxy unless we are in normal mode
@@ -242,4 +246,60 @@ func (c *conn) run(ctx context.Context) (err error) {
 			return
 		}
 	}
+}
+
+// route routes to common.Router and measures all the result
+func (c *conn) route(h common.Handler, ctx context.Context, reqHeader *wire.MsgHeader, reqBody wire.MsgBody,
+) (resHeader *wire.MsgHeader, resBody wire.MsgBody, closeConn bool) {
+	var err error
+	requests := c.m.requests.MustCurryWith(prometheus.Labels{"opcode": reqHeader.OpCode.String()})
+	var command string
+	switch reqHeader.OpCode {
+	case wire.OP_MSG:
+		msg := reqBody.(*wire.OpMsg)
+		var document *types.Document
+		document, err = msg.Document()
+		command = document.Command()
+
+	case wire.OP_QUERY:
+		query := reqBody.(*wire.OpQuery)
+		command = query.Query.Command()
+
+	case wire.OP_REPLY:
+		fallthrough
+	case wire.OP_UPDATE:
+		fallthrough
+	case wire.OP_INSERT:
+		fallthrough
+	case wire.OP_GET_BY_OID:
+		fallthrough
+	case wire.OP_GET_MORE:
+		fallthrough
+	case wire.OP_DELETE:
+		fallthrough
+	case wire.OP_KILL_CURSORS:
+		fallthrough
+	case wire.OP_COMPRESSED:
+		fallthrough
+	default:
+		err = lazyerrors.Errorf("unexpected OpCode %s", reqHeader.OpCode)
+	}
+	requests.WithLabelValues(command).Inc()
+
+	var result *string
+	resHeader, resBody, closeConn, result = common.Route(c.h, ctx, reqHeader, reqBody)
+
+	if result != nil && *result == "unexpected" {
+		c.l.Error("Handler error for unexpected response opcode",
+			zap.Error(err), zap.Stringer("opcode", resHeader.OpCode),
+		)
+	}
+	c.m.responses.WithLabelValues(resHeader.OpCode.String(), command, *result).Inc()
+
+	// do not spend time dumping if we are not going to log it
+	if c.l.Desugar().Core().Enabled(zap.DebugLevel) {
+		c.l.Debugf("Response header: %s", resHeader)
+		c.l.Debugf("Response message:\n%s\n\n\n", resBody)
+	}
+	return
 }
