+++ internal/handlers/common/handler.go
@@ -17,16 +17,11 @@ package common
 import (
 	"context"
 
-	"github.com/prometheus/client_golang/prometheus"
-
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
 
 // Handler interface represents common commands handlers.
 type Handler interface {
-	// Handle handles the message.
-	Handle(context.Context, *wire.MsgHeader, wire.MsgBody) (*wire.MsgHeader, wire.MsgBody, bool)
-
 	// MsgListCommands returns a list of currently supported commands.
 	MsgListCommands(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error)
 
@@ -113,12 +108,6 @@ type Handler interface {
 
 	// CmdQuery runs query operation command.
 	CmdQuery(ctx context.Context, query *wire.OpQuery) (*wire.OpReply, error)
-
-	// Collect implements prometheus.Collector.
-	Collect(chan<- prometheus.Metric)
-
-	// Describe implements prometheus.Collector.
-	Describe(ch chan<- *prometheus.Desc)
 
 	// Close prepares handler for graceful shutdown: closes connections, channels etc.
 	Close()
