+++ internal/handlers/dummy/dummy.go
@@ -18,8 +18,6 @@ package dummy
 import (
 	"context"
 
-	"github.com/prometheus/client_golang/prometheus"
-
 	"github.com/FerretDB/FerretDB/internal/handlers/common"
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
@@ -30,11 +28,6 @@ type Handler struct{}
 // New returns a new handler.
 func New() common.Handler {
 	return new(Handler)
-}
-
-// Handle handles the message.
-func (h *Handler) Handle(context.Context, *wire.MsgHeader, wire.MsgBody) (*wire.MsgHeader, wire.MsgBody, bool) {
-	return nil, nil, false
 }
 
 // MsgListCommands returns information about the currently supported commands.
@@ -181,12 +174,6 @@ func (h *Handler) MsgDebugPanic(ctx context.Context, msg *wire.OpMsg) (*wire.OpM
 func (h *Handler) CmdQuery(ctx context.Context, query *wire.OpQuery) (*wire.OpReply, error) {
 	return nil, common.NewErrorMsg(common.ErrNotImplemented, "I'm a dummy, not a handler")
 }
-
-// Collect implements prometheus.Collector.
-func (h *Handler) Collect(chan<- prometheus.Metric) {}
-
-// Describe implements prometheus.Collector.
-func (h *Handler) Describe(ch chan<- *prometheus.Desc) {}
 
 // Close prepares handler for graceful shutdown: closes connections, channels etc.
 func (h *Handler) Close() {}
